refactor(guardrails): split guardrail.ts into 6 modules (Issue #150)
Split monolithic guardrail.ts (1944 lines) into modular architecture:
- guardrail-patterns.ts (225L): constants, pure utilities
- guardrail-context.ts (343L): GuardrailContext type, createContext()
- guardrail-review.ts (198L): auto-review pipeline, dual review gate
- guardrail-git.ts (432L): branch hygiene, merge gates, CI checks
- guardrail-access.ts (296L): file access control, delegation, tracking
- guardrail.ts (568L): entry point, hook registration

Architecture: Context Object + Handler Factories pattern.
Dependency DAG: patterns <- context <- {review, git, access} <- entry.
All files under 800-line limit. 34/34 scenario tests pass.
